refactor: isolate function to build main application instance
